run general code cleanup

+++ src/com/owncloud/android/widgets/ActionEditText.java
@@ -73,7 +73,7 @@ public class ActionEditText extends EditText {
         getDrawingRect(mButtonRect);
         mButtonRect.top += 10;
         mButtonRect.bottom -= 10;
-        mButtonRect.left = (int) (getWidth() - mTextBounds.width() - 18);
+        mButtonRect.left = getWidth() - mTextBounds.width() - 18;
         mButtonRect.right = getWidth() - 10;
         btn_rect = mButtonRect;
